requirements.txt
google-genai>=1.10.0
gradio>=5.0.0
